Numpad puzzle (description for more info)
Added numpad puzzle and the required VR elements in Leo_Test (scene). Functionality tested in normal mouse interaction via Canvas buttons and deemed working. Needs testing in VR, and hooking up the puzzle to the door. Might need custom assets for the new keypad.
NOTE: not sure if the buttons are too big to look good and too small to work in VR.